updated devise + controllers + omniauth
Instead of asking OmniAuth to retrieve all the information from
Facebook, i’m using Omniauth purely to source the Facebook ID and
attach it to a new user account.

As a result, I’ve had to remove the password confirmation on my edit
page as generating a new user account via OmniAuth does not actually
create a password.

I’ve also removed the validations from the User model which initially
required birthday, phone number to be filled out to become a user.
Instead, I’ve added a method which redirects you to the edit user
registration page if this is the first time you’ve signed in (aka. new
user).

TO-DO: Create a separate page to edit password which DOES require a
password confirmation.

TO-DO: Create a way to display which Facebook profile is linked WITHOUT
showing the Facebook ID (and therefore, allowing the user to edit).